Fixed cleanMapAtSave usage, dynamic broadcast for next save
cleanMap was commented out. Removed the comment, and defaulted the clean
to false. User can toggle to true if they want the map to clean. Also
adjusted the broadcast to tell when the next save is based on the
interval of the global event instead of a static number, incase someone
wants to make it faster, they dont have to edit the script too.