Fixed open channels. Formatting adjustments.
Made it so it opens rookgaard advertising channel if the players town is
Rookgaard or Dawnport instead of basing it off vocations and level since
it is possible to have vocations in Dawnport. I also made it compare
town names so the ID's can change with map updates, but the names should
remain the same.